% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_my_stanfolder}
\alias{get_my_stanfolder}
\title{Check where your stan models are stored.}
\usage{
get_my_stanfolder()
}
\description{
Add .stan files to this folder to make them available across projects.
}
