% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_stanmodel}
\alias{load_stanmodel}
\title{Load a compiled stan model, or compile it if necessary}
\usage{
load_stanmodel(model_name, force_recompile = FALSE, avoid_recompile = FALSE)
}
\arguments{
\item{model_name}{The name of the model to load}

\item{force_recompile}{Recompile the model even if it already exists?
Default FALSE
Recompilation will only happen if the .stan file has changed.}

\item{avoid_recompile}{If the model doesn't exist, don't bother compiling it?
Default FALSE}
}
\description{

}
\details{
Stan model code is found in ./inst/stan/
(relative to whatever directory you've cloned this repository to).
Compiled models are saved to ~/.mystanmodels/ (where ~ is your home directory).
}
\examples{
\dontrun{
model = lab_stanmodel('simple_lm')
N = 100
x = rnorm(N, 0, 1)
y = 10 + x * 2 + rnorm(N, 0, 1)
data = list(x=x, y=y, N=N)
posterior_samples = rstan::sampling(model, data=data, iter=2000)
est_map = rstan::optimizing(model, data=data)
est_vb = rstan::vb(model, data=data)
print(posterior_samples)
print(est_map$par)
print(est_vb)
}
}
